+++ core_planning/twist_filter/include/accel_limiter.hpp
@@ -1,38 +1,17 @@
 #pragma once
-/*
- * Copyright (C) 2018-2021 LEIDOS.
- *
- * Licensed under the Apache License, Version 2.0 (the "License"); you may not
- * use this file except in compliance with the License. You may obtain a copy of
- * the License at
- *
- * http://www.apache.org/licenses/LICENSE-2.0
- *
- * Unless required by applicable law or agreed to in writing, software
- * distributed under the License is distributed on an "AS IS" BASIS, WITHOUT
- * WARRANTIES OR CONDITIONS OF ANY KIND, either express or implied. See the
- * License for the specific language governing permissions and limitations under
- * the License.
- */
 
-#include <ros/ros.h>
-#include <geometry_msgs/TwistStamped.h>
-#include <autoware_msgs/ControlCommandStamped.h>
+#include <rclcpp/rclcpp.hpp>
+#include <autoware_msgs/msg/control_command_stamped.hpp>
+#include <geometry_msgs/msg/twist_stamped.hpp>
 #include <hardcoded_params/control_limits/control_limits.h>
 
-namespace twist_filter {
+namespace twist_filter{
 
 constexpr double MAX_LONGITUDINAL_ACCEL_HARDCODED_LIMIT_M_S_2 = hardcoded_params::control_limits::MAX_LONGITUDINAL_ACCEL_MPS2;
 
-/**
- * \brief Class responsible for encapsulating the state related to acceleration
- * limiting of longitudinal speed commands. Keeps track of the previously received
- * message's speeds and times to compute acceleration against the new message's
- * value.
- */
-class LongitudinalAccelLimiter 
+class LongitudinalAccelLimiter
 {
-    public:
+    public: 
         /**
          * Default constructor, not recommended for use.
          */
@@ -41,6 +20,7 @@ class LongitudinalAccelLimiter
             _initialized(false),
             _prev_v(0.0),
             _prev_t(0.0) {};
+
 
         /**
          * \brief Constructor for LongitudinalAccelLimiter
@@ -52,7 +32,7 @@ class LongitudinalAccelLimiter
             _accel_limit(accel_limit),
             _initialized(false),
             _prev_v(0.0),
-            _prev_t(0.0) {};
+            _prev_t(0.0) {};  
 
         /**
          * Limit the longitudinal acceleration found in the input ControlCommandStamped
@@ -61,24 +41,24 @@ class LongitudinalAccelLimiter
          * \return A copy of the message with the longitudinal accel limited 
          * based on params or hardcoded limit
          */
-        autoware_msgs::ControlCommandStamped
-            longitudinalAccelLimitCtrl(const autoware_msgs::ControlCommandStamped& msg);
+        autoware_msgs::msg::ControlCommandStamped
+            longitudinalAccelLimitCtrl(const autoware_msgs::msg::ControlCommandStamped& msg);
 
-        /**
+                /**
          * Limit the longitudinal acceleration found in the input TwistStamped
          * 
          * \param msg The message to be evaluated
          * \return A copy of the message with the longitudinal accel limited 
          * based on params or hardcoded limit
          */
-        geometry_msgs::TwistStamped
-            longitudinalAccelLimitTwist(const geometry_msgs::TwistStamped& msg);
+        geometry_msgs::msg::TwistStamped
+            longitudinalAccelLimitTwist(const geometry_msgs::msg::TwistStamped& msg);
 
     private:
         double _accel_limit;
         bool _initialized;
         double _prev_v;
-        ros::Time _prev_t;
+        rclcpp::Time _prev_t; 
 };
 
 }
