--- core_planning/twist_filter/launch/twist_filter.launch
@@ -1,28 +1,0 @@
-<?xml version="1.0"?>
-<launch>
-  <arg name="loop_rate" default="30.0" />
-  <arg name="use_decision_maker" default="false" />
-  <arg name="lateral_accel_limit" default="5.0" />
-  <arg name="lateral_jerk_limit" default="5.0" />
-  <arg name="lowpass_gain_linear_x" default="0.0" />
-  <arg name="lowpass_gain_angular_z" default="0.0" />
-  <arg name="lowpass_gain_steering_angle" default="0.0" />
-  <arg name="longitudinal_velocity_limit" default="80.0" /> <!--mph-->
-  <arg name="longitudinal_accel_limit" default="3.5" />
-
-  <!-- rosrun waypoint_follower twist_filter -->
-  <node pkg="twist_filter" type="twist_filter" name="twist_filter" output="log">
-    <param name="longitudinal_velocity_limit" value="$(arg longitudinal_velocity_limit)" />
-    <param name="longitudinal_accel_limit" value="$(arg longitudinal_accel_limit)" />
-    <param name="lateral_accel_limit" value="$(arg lateral_accel_limit)" />
-    <param name="lateral_jerk_limit" value="$(arg lateral_jerk_limit)" />
-    <param name="lowpass_gain_linear_x" value="$(arg lowpass_gain_linear_x)" />
-    <param name="lowpass_gain_angular_z" value="$(arg lowpass_gain_angular_z)" />
-    <param name="lowpass_gain_steering_angle" value="$(arg lowpass_gain_steering_angle)" />
-  </node>
-
-  <node pkg="twist_gate" type="twist_gate" name="twist_gate" output="log">
-    <param name="loop_rate" value="$(arg loop_rate)" />
-    <param name="use_decision_maker" value="$(arg use_decision_maker)" />
-  </node>
-</launch>
