+++ core_planning/twist_filter/launch/twist_filter_launch.py
@@ -33,35 +33,34 @@ def generate_launch_description():
 
     # Declare the log_level launch argument
     log_level = LaunchConfiguration('log_level')
-    declare_log_level_arg = DeclareLaunchArgument(
-        name ='log_level', default_value='WARN')
+    declare_log_level_arg = DeclareLaunchArgument(name ='log_level', default_value='WARN')
 
     loop_rate = LaunchConfiguration('loop_rate')
-    declare_loop_rate = DeclareLaunchArgument(name='loop_rate', default_value=30)
+    declare_loop_rate = DeclareLaunchArgument(name='loop_rate', default_value='30.0')
 
     use_decision_maker = LaunchConfiguration('use_decision_maker')
-    declare_use_decision_maker = DeclareLaunchArgument(name='use_decision_maker', default_value=False)
+    declare_use_decision_maker = DeclareLaunchArgument(name='use_decision_maker', default_value='False')
 
     lateral_accel_limit = LaunchConfiguration('lateral_accel_limit')
-    declare_lateral_accel_limit = DeclareLaunchArgument(name = 'lateral_accel_limit', default_value=5.0)
+    declare_lateral_accel_limit = DeclareLaunchArgument(name = 'lateral_accel_limit', default_value='5.0')
 
     lateral_jerk_limit = LaunchConfiguration('lateral_jerk_limit')
-    declare_lateral_jerk_limit = DeclareLaunchArgument(name = 'lateral_jerk_limit', default_value=5.0)
+    declare_lateral_jerk_limit = DeclareLaunchArgument(name = 'lateral_jerk_limit', default_value='5.0')
 
     lowpass_gain_linear_x = LaunchConfiguration('lowpass_gain_linear_x')
-    declare_lowpass_gain_linear_x = DeclareLaunchArgument(name = 'lowpass_gain_linear_x', default_value=0.0)
+    declare_lowpass_gain_linear_x = DeclareLaunchArgument(name = 'lowpass_gain_linear_x', default_value='0.0')
 
     lowpass_gain_angular_z = LaunchConfiguration('lowpass_gain_angular_z')
-    declare_lowpass_gain_angular_z = DeclareLaunchArgument(name='lowpass_gain_angular_z', default_value=0.0)
+    declare_lowpass_gain_angular_z = DeclareLaunchArgument(name='lowpass_gain_angular_z', default_value='0.0')
 
     lowpass_gain_steering_angle = LaunchConfiguration('lowpass_gain_steering_angle')
-    declare_lowpass_gain_steering_angle = DeclareLaunchArgument(name='lowpass_gain_steering_angle', default_value=0.0)
+    declare_lowpass_gain_steering_angle = DeclareLaunchArgument(name='lowpass_gain_steering_angle', default_value='0.0')
 
     longitudinal_velocity_limit = LaunchConfiguration('longitudinal_velocity_limit')
-    declare_longitudinal_velocity_limit = DeclareLaunchArgument(name = 'longitudinal_velocity_limit', default_value=80.0) #mph
+    declare_longitudinal_velocity_limit = DeclareLaunchArgument(name = 'longitudinal_velocity_limit', default_value='80.0') #mph
 
     longitudinal_accel_limit = LaunchConfiguration('longitudinal_accel_limit')
-    declare_longitudinal_accel_limit = DeclareLaunchArgument(name= 'longitudinal_accel_limit', default_value = 3.5)
+    declare_longitudinal_accel_limit = DeclareLaunchArgument(name= 'longitudinal_accel_limit', default_value='3.5')
 
         
     # Launch node(s) in a carma container to allow logging to be configured
@@ -81,12 +80,7 @@ def generate_launch_description():
                         {'use_intra_process_comms': True},
                         {'--log-level' : log_level },
                     ],
-                    remappings = [
-
-                    ],
                     parameters = [
-                        {'loop_rate': loop_rate},
-                        {'use_decision_maker' : use_decision_maker},
                         {'lateral_accel_limit' : lateral_accel_limit},
                         {'lateral_jerk_limit' : lateral_jerk_limit},
                         {'lowpass_gain_linear_x' : lowpass_gain_linear_x},
@@ -96,10 +90,30 @@ def generate_launch_description():
                         {'longitudinal_accel_limit' : longitudinal_accel_limit}
                     ]
             ),
+            ComposableNode(
+                package='twist_gate',
+                plugin='TwistGate',
+                name='twist_gate_node',
+                extra_arguments=[
+                    {'use_intra_process_comms': True},
+                    {'--log-level' : log_level },
+                    {'loop_rate': loop_rate},
+                    {'use_decision_maker' : use_decision_maker}
+                ]
+            )
         ]
     )
 
     return LaunchDescription([
         declare_log_level_arg,
+        declare_loop_rate,
+        declare_use_decision_maker,
+        declare_lateral_accel_limit,
+        declare_lateral_jerk_limit,
+        declare_lowpass_gain_linear_x,
+        declare_lowpass_gain_angular_z,
+        declare_lowpass_gain_steering_angle,
+        declare_longitudinal_velocity_limit,
+        declare_longitudinal_accel_limit,
         container
     ])
