+++ core_planning/twist_filter/src/accel_limit.cpp
@@ -1,5 +1,5 @@
 /*
- * Copyright (C) 2018-2021 LEIDOS.
+ * Copyright (C) 2022 LEIDOS.
  *
  * Licensed under the Apache License, Version 2.0 (the "License"); you may not
  * use this file except in compliance with the License. You may obtain a copy of
@@ -18,23 +18,23 @@
 #include <limits>
 #include <boost/math/special_functions/sign.hpp>
 
-namespace twist_filter {
+namespace twist_filter{
 
-autoware_msgs::ControlCommandStamped
-LongitudinalAccelLimiter::longitudinalAccelLimitCtrl(const autoware_msgs::ControlCommandStamped& msg) {
+autoware_msgs::msg::ControlCommandStamped
+LongitudinalAccelLimiter::longitudinalAccelLimitCtrl(const autoware_msgs::msg::ControlCommandStamped& msg) {
     if (!_initialized) {
         // Record current command and then return unmodified
         _prev_v = msg.cmd.linear_velocity;
-        _prev_t = msg.header.stamp;
+        _prev_t = rclcpp::Time(msg.header.stamp);
         _initialized = true;
 
         return msg;
     }
 
-    autoware_msgs::ControlCommandStamped out{msg};
+    autoware_msgs::msg::ControlCommandStamped out{msg};
     // Else we've initialized previously, let's compare.
     double delta_v = msg.cmd.linear_velocity - _prev_v;
-    double delta_t = (msg.header.stamp - _prev_t).toSec();
+    double delta_t = (rclcpp::Time(msg.header.stamp) - _prev_t).seconds();
     double accel = delta_v / delta_t;
 
     if (std::abs(accel) > _accel_limit) {
@@ -44,26 +44,26 @@ LongitudinalAccelLimiter::longitudinalAccelLimitCtrl(const autoware_msgs::Contro
 
     // Update filter for next timestep
     _prev_v = out.cmd.linear_velocity;
-    _prev_t = out.header.stamp;
+    _prev_t = rclcpp::Time(out.header.stamp);
 
     return out;
 }
 
-geometry_msgs::TwistStamped
-LongitudinalAccelLimiter::longitudinalAccelLimitTwist(const geometry_msgs::TwistStamped& msg) {
+geometry_msgs::msg::TwistStamped
+LongitudinalAccelLimiter::longitudinalAccelLimitTwist(const geometry_msgs::msg::TwistStamped& msg) {
     if (!_initialized) {
         // Record current command and then return unmodified
         _prev_v = msg.twist.linear.x;
-        _prev_t = msg.header.stamp;
+        _prev_t = rclcpp::Time(msg.header.stamp);
         _initialized = true;
 
         return msg;
     }
 
-    geometry_msgs::TwistStamped out{msg};
+    geometry_msgs::msg::TwistStamped out{msg};
     // Else we've initialized previously, let's compare.
     double delta_v = msg.twist.linear.x - _prev_v;
-    double delta_t = (msg.header.stamp - _prev_t).toSec();
+    double delta_t = (rclcpp::Time(msg.header.stamp) - _prev_t).seconds();
     double accel = delta_v / delta_t;
 
     if (std::abs(accel) > _accel_limit) {
@@ -72,9 +72,11 @@ LongitudinalAccelLimiter::longitudinalAccelLimitTwist(const geometry_msgs::Twist
 
     // Update filter for next timestep
     _prev_v = out.twist.linear.x;
-    _prev_t = out.header.stamp;
+    _prev_t = rclcpp::Time(out.header.stamp);
 
     return out;
 }
+
+
 
 }
