+++ core_planning/twist_filter/src/twist_filter.cpp
@@ -25,8 +25,70 @@
 #include "twist_filter.hpp"
 #include "velocity_limit.hpp"
 
+
 namespace twist_filter
 {
+
+namespace std_ph = std::placeholders;
+
+TwistFilter::TwistFilter(const rclcpp::NodeOptions &options)
+    : carma_ros2_utils::CarmaLifecycleNode(options)
+{
+
+  // Declare parameters
+  wheel_base_ = declare_parameter<double>("vehicle_info/wheel_base", wheel_base_);
+  longitudinal_velocity_limit_ = declare_parameter<double>("longitudinal_velocity_limit", longitudinal_velocity_limit_);
+  longitudinal_accel_limit_ = declare_parameter<double>("longitudinal_accel_limit", longitudinal_accel_limit_);
+  lateral_accel_limit_ = declare_parameter<double>("lateral_accel_limit", lateral_accel_limit_);
+  lateral_jerk_limit_ = declare_parameter<double>("lateral_jerk_limit", lateral_jerk_limit_);
+  lowpass_gain_linear_x_ = declare_parameter<double>("lowpass_gain_linear_x", lowpass_gain_linear_x_);
+  lowpass_gain_angular_z_ = declare_parameter<double>("lowpass_gain_angular_x", lowpass_gain_angular_z_);
+  lowpass_gain_steering_angle_ = declare_parameter<double>("lowpass_gain_steering_angle", lowpass_gain_steering_angle_);
+}
+
+carma_ros2_utils::CallbackReturn TwistFilter::handle_on_configure(const rclcpp_lifecycle::State &prev_state)
+{
+  // get parameters
+  get_parameter<double>("vehicle_info/wheel_base", wheel_base_);
+  // parameters on private handles
+  get_parameter<double>("~/longitudinal_velocity_limit", longitudinal_velocity_limit_);
+  longitudinal_velocity_limit_ = longitudinal_velocity_limit_ * 0.44704;
+  
+  get_parameter<double>("~/longitudinal_accel_limit", longitudinal_accel_limit_);
+  _lon_accel_limiter = LongitudinalAccelLimiter{
+    std::min(longitudinal_accel_limit_, MAX_LONGITUDINAL_ACCEL_HARDCODED_LIMIT_M_S_2)};
+  
+  get_parameter<double>("~/lateral_accel_limit", lateral_accel_limit_);
+  get_parameter<double>("~/lateral_jerk_limit", lateral_jerk_limit_);
+  get_parameter<double>("~/lowpass_gain_linear_x", lowpass_gain_linear_x_);
+  get_parameter<double>("~/lowpass_gain_angular_x", lowpass_gain_angular_z_);
+  get_parameter<double>("~/lowpass_gain_steering_angle", lowpass_gain_steering_angle_);
+  
+
+  //Setup subscribers
+  twist_sub_ = create_subscription<geometry_msgs::msg::TwistStamped>("twist_raw", 1, std::bind(&TwistFilter::TwistCmdCallback, this, std_ph::_1));
+  ctrl_sub_ = create_subscription<autoware_msgs::msg::ControlCommandStamped>("ctrl_raw", 1, std::bind(&TwistFilter::CtrlCmdCallback, this, std_ph::_1));
+  config_sub_ = create_subscription<autoware_config_msgs::msg::ConfigTwistFilter>("config/twist_filter", 10, std::bind(&TwistFilter::configCallback, this, std_ph::_1));
+
+  //Setup publishers
+  twist_pub_ = create_publisher<geometry_msgs::msg::TwistStamped>("twist_cmd", 5);
+  ctrl_pub_ = create_publisher<autoware_msgs::msg::ControlCommandStamped>("ctrl_cmd", 5);
+  
+  // Publishers on private handles
+  twist_lacc_limit_debug_pub_ = create_publisher<std_msgs::msg::Float32>("~/limitation_debug/twist/lateral_accel", 5);
+  twist_ljerk_limit_debug_pub_ = create_publisher<std_msgs::msg::Float32>("~/limitation_debug/twist/lateral_jerk", 5);
+  ctrl_lacc_limit_debug_pub_ = create_publisher<std_msgs::msg::Float32>("~/limitation_debug/ctrl/lateral_accel", 5);
+  ctrl_ljerk_limit_debug_pub_ = create_publisher<std_msgs::msg::Float32>("~/limitation_debug/ctrl/lateral_jerk", 5);
+  twist_lacc_result_pub_ = create_publisher<std_msgs::msg::Float32>("~/result/twist/lateral_accel", 5);
+  twist_ljerk_result_pub_ = create_publisher<std_msgs::msg::Float32>("~/result/twist/lateral_jerk", 5);
+  ctrl_lacc_result_pub_ = create_publisher<std_msgs::msg::Float32>("~/result/ctrl/lateral_accel", 5);
+  ctrl_ljerk_result_pub_ = create_publisher<std_msgs::msg::Float32>("~/result/ctrl/lateral_jerk", 5);
+  
+
+  return CallbackReturn::SUCCESS;
+
+}
+
 
 auto lowpass_filter()
 {
@@ -77,7 +139,7 @@ boost::optional<double> TwistFilter::calcLjerkWithSteeringAngle(
 }
 
 void TwistFilter::publishLateralResultsWithTwist(
-  const geometry_msgs::TwistStamped& msg) const
+  const geometry_msgs::msg::TwistStamped& msg) const
 {
   const double lv = msg.twist.linear.x;
   const double az = msg.twist.angular.z;
@@ -87,15 +149,15 @@ void TwistFilter::publishLateralResultsWithTwist(
   {
     return;
   }
-  std_msgs::Float32 lacc_msg, ljerk_msg;
+  std_msgs::msg::Float32 lacc_msg, ljerk_msg;
   lacc_msg.data = lacc.get();
   ljerk_msg.data = ljerk.get();
-  twist_lacc_result_pub_.publish(lacc_msg);
-  twist_ljerk_result_pub_.publish(ljerk_msg);
+  twist_lacc_result_pub_->publish(lacc_msg);
+  twist_ljerk_result_pub_->publish(ljerk_msg);
 }
 
 void TwistFilter::publishLateralResultsWithCtrl(
-  const autoware_msgs::ControlCommandStamped& msg) const
+  const autoware_msgs::msg::ControlCommandStamped& msg) const
 {
   const double lv = msg.cmd.linear_velocity;
   const double sa = msg.cmd.steering_angle;
@@ -105,75 +167,42 @@ void TwistFilter::publishLateralResultsWithCtrl(
   {
     return;
   }
-  std_msgs::Float32 lacc_msg, ljerk_msg;
+  std_msgs::msg::Float32 lacc_msg, ljerk_msg;
   lacc_msg.data = lacc.get();
   ljerk_msg.data = ljerk.get();
-  ctrl_lacc_result_pub_.publish(lacc_msg);
-  ctrl_ljerk_result_pub_.publish(ljerk_msg);
+  ctrl_lacc_result_pub_->publish(lacc_msg);
+  ctrl_ljerk_result_pub_->publish(ljerk_msg);
 }
 
-void TwistFilter::checkTwist(const geometry_msgs::TwistStamped& msg)
+void TwistFilter::checkTwist(const geometry_msgs::msg::TwistStamped& msg)
 {
   const double lv = msg.twist.linear.x;
   const double az = msg.twist.angular.z;
   const auto lacc = calcLaccWithAngularZ(lv, az);
   const auto ljerk = calcLjerkWithAngularZ(lv, az);
 
-  // Check Longitudinal velocity vs limiter
-  health_checker_.CHECK_MAX_VALUE("twist_longitudinal_v_high",
-    lv, longitudinal_velocity_limit_ * 0.9, longitudinal_velocity_limit_, MAX_LONGITUDINAL_VELOCITY_HARDCODED_LIMIT_M_S,
-    "longitudinal velocity is too high in twist_filtering");
-
-  if (lacc)
-  {
-    health_checker_.CHECK_MAX_VALUE("twist_lateral_accel_high",
-      lacc.get(), lateral_accel_limit_, 2 * lateral_accel_limit_, DBL_MAX,
-      "lateral_accel is too high in twist filtering");
-  }
-  if (ljerk)
-  {
-    health_checker_.CHECK_MAX_VALUE("twist_lateral_jerk_high",
-      lacc.get(), lateral_jerk_limit_, 2 * lateral_jerk_limit_, DBL_MAX,
-      "lateral_jerk is too high in twist filtering");
-  }
 }
 
-void TwistFilter::checkCtrl(const autoware_msgs::ControlCommandStamped& msg)
+void TwistFilter::checkCtrl(const autoware_msgs::msg::ControlCommandStamped& msg)
 {
   const double lv = msg.cmd.linear_velocity;
   const double sa = msg.cmd.steering_angle;
   const auto lacc = calcLaccWithSteeringAngle(lv, sa);
   const auto ljerk = calcLjerkWithSteeringAngle(lv, sa);
 
-  // Check Longitudinal velocity vs limiter
-  health_checker_.CHECK_MAX_VALUE("ctrl_longitudinal_v_high",
-    lv, longitudinal_velocity_limit_ * 0.9, longitudinal_velocity_limit_, MAX_LONGITUDINAL_VELOCITY_HARDCODED_LIMIT_M_S,
-    "longitudinal velocity is too high in ctrl_filtering");
-
-  if (lacc)
-  {
-    health_checker_.CHECK_MAX_VALUE("ctrl_lateral_accel_high",
-      lacc.get(), lateral_accel_limit_, 3 * lateral_accel_limit_, DBL_MAX,
-      "lateral_accel is too high in ctrl filtering");
-  }
-  if (ljerk)
-  {
-    health_checker_.CHECK_MAX_VALUE("ctrl_lateral_jerk_high",
-      lacc.get(), lateral_jerk_limit_, 3 * lateral_jerk_limit_, DBL_MAX,
-      "lateral_jerk is too high in ctrl filtering");
-  }
 }
 
-geometry_msgs::TwistStamped
-  TwistFilter::lateralLimitTwist(const geometry_msgs::TwistStamped& msg)
+
+geometry_msgs::msg::TwistStamped
+  TwistFilter::lateralLimitTwist(const geometry_msgs::msg::TwistStamped& msg)
 {
   static bool init = false;
 
-  geometry_msgs::TwistStamped ts;
+  geometry_msgs::msg::TwistStamped ts;
   ts = msg;
 
-  ros::Time t = msg.header.stamp;
-  az_prev_.dt = (t - az_prev_.time).toSec();
+  rclcpp::Time t = rclcpp::Time(msg.header.stamp);
+  az_prev_.dt = (t - az_prev_.time).seconds();
   const double lv = msg.twist.linear.x;
   double az = msg.twist.angular.z;
 
@@ -192,8 +221,8 @@ geometry_msgs::TwistStamped
   {
     double sgn = lacc / std::fabs(lacc);
     double az_max = sgn * (lateral_accel_limit_) / lv;
-    ROS_WARN_THROTTLE(1,
-      "Limit angular velocity by lateral acceleration: %f -> %f", az, az_max);
+    auto& clk = *this->get_clock();
+    RCLCPP_WARN_THROTTLE(get_logger(), clk, 1, "Limit angular velocity by lateral acceleration: %f -> %f", az, az_max);
     az = az_max;
   }
 
@@ -205,8 +234,8 @@ geometry_msgs::TwistStamped
     double sgn = ljerk / std::fabs(ljerk);
     double az_max =
       az_prev_.val + (sgn * lateral_jerk_limit_ / lv) * az_prev_.dt;
-    ROS_WARN_THROTTLE(1,
-      "Limit angular velocity by lateral jerk: %f -> %f", az, az_max);
+    auto& clk = *this->get_clock();
+    RCLCPP_WARN_THROTTLE(get_logger(), clk, 1,"Limit angular velocity by lateral jerk: %f -> %f", az, az_max);
     az = az_max;
   }
 
@@ -218,22 +247,22 @@ geometry_msgs::TwistStamped
   ljerk = calcLjerkWithAngularZ(lv, az).get();
 
   // for debug
-  std_msgs::Float32 lacc_msg, ljerk_msg;
+  std_msgs::msg::Float32 lacc_msg, ljerk_msg;
   lacc_msg.data = lacc;
   ljerk_msg.data = ljerk;
-  twist_lacc_limit_debug_pub_.publish(lacc_msg);
-  twist_ljerk_limit_debug_pub_.publish(ljerk_msg);
+  twist_lacc_limit_debug_pub_->publish(lacc_msg);
+  twist_ljerk_limit_debug_pub_->publish(ljerk_msg);
 
   return ts;
 }
 
-geometry_msgs::TwistStamped
-  TwistFilter::smoothTwist(const geometry_msgs::TwistStamped& msg)
+geometry_msgs::msg::TwistStamped
+  TwistFilter::smoothTwist(const geometry_msgs::msg::TwistStamped& msg)
 {
   static auto lp_lx = lowpass_filter();
   static auto lp_az = lowpass_filter();
 
-  geometry_msgs::TwistStamped ts;
+  geometry_msgs::msg::TwistStamped ts;
   ts = msg;
 
   // apply lowpass filter to linear_x / angular_z
@@ -243,16 +272,16 @@ geometry_msgs::TwistStamped
   return ts;
 }
 
-autoware_msgs::ControlCommandStamped
-  TwistFilter::lateralLimitCtrl(const autoware_msgs::ControlCommandStamped& msg)
+autoware_msgs::msg::ControlCommandStamped
+  TwistFilter::lateralLimitCtrl(const autoware_msgs::msg::ControlCommandStamped& msg)
 {
   static bool init = false;
 
-  autoware_msgs::ControlCommandStamped ccs;
+  autoware_msgs::msg::ControlCommandStamped ccs;
   ccs = msg;
 
-  ros::Time t = msg.header.stamp;
-  sa_prev_.dt = (t - sa_prev_.time).toSec();
+  rclcpp::Time t = rclcpp::Time(msg.header.stamp);
+  sa_prev_.dt = (t - sa_prev_.time).seconds();
   const double lv = msg.cmd.linear_velocity;
   double sa = msg.cmd.steering_angle;
 
@@ -272,8 +301,9 @@ autoware_msgs::ControlCommandStamped
     double sgn = lacc / std::fabs(lacc);
     double sa_max =
       std::atan(sgn * lateral_accel_limit_ * wheel_base_ / (lv * lv));
-    ROS_WARN_THROTTLE(1,
-      "Limit steering angle by lateral acceleration: %f -> %f", sa, sa_max);
+    auto& clk = *this->get_clock();
+    RCLCPP_WARN_THROTTLE(get_logger(), clk, 1,
+        "Limit steering angle by lateral acceleration: %f -> %f", sa, sa_max);
     sa = sa_max;
   }
 
@@ -285,12 +315,13 @@ autoware_msgs::ControlCommandStamped
     double sgn = ljerk / std::fabs(ljerk);
     double sa_max = std::atan(std::tan(sa_prev_.val) +
       sgn * (lateral_jerk_limit_ * wheel_base_ / (lv * lv)) * sa_prev_.dt);
-    ROS_WARN_THROTTLE(1,
-      "Limit steering angle by lateral jerk: %f -> %f", sa, sa_max);
+    auto& clk = *this->get_clock();
+    RCLCPP_WARN_THROTTLE(get_logger(), clk, 1,
+     "Limit steering angle by lateral jerk: %f -> %f",sa, sa_max);
     sa = sa_max;
   }
 
-  // update by lateral limitaion
+  // update by lateral limitation
   ccs.cmd.steering_angle = sa;
 
   // update lateral acceleration/jerk
@@ -298,22 +329,22 @@ autoware_msgs::ControlCommandStamped
   ljerk = calcLjerkWithSteeringAngle(lv, sa).get();
 
   // for debug
-  std_msgs::Float32 lacc_msg, ljerk_msg;
+  std_msgs::msg::Float32 lacc_msg, ljerk_msg;
   lacc_msg.data = lacc;
   ljerk_msg.data = ljerk;
-  ctrl_lacc_limit_debug_pub_.publish(lacc_msg);
-  ctrl_ljerk_limit_debug_pub_.publish(ljerk_msg);
+  ctrl_lacc_limit_debug_pub_->publish(lacc_msg);
+  ctrl_ljerk_limit_debug_pub_->publish(ljerk_msg);
 
   return ccs;
 }
 
-autoware_msgs::ControlCommandStamped
-  TwistFilter::smoothCtrl(const autoware_msgs::ControlCommandStamped& msg)
+autoware_msgs::msg::ControlCommandStamped
+  TwistFilter::smoothCtrl(const autoware_msgs::msg::ControlCommandStamped& msg)
 {
   static auto lp_lx = lowpass_filter();
   static auto lp_sa = lowpass_filter();
 
-  autoware_msgs::ControlCommandStamped ccs;
+  autoware_msgs::msg::ControlCommandStamped ccs;
   ccs = msg;
 
   // apply lowpass filter to linear_x / steering_angle
@@ -326,7 +357,7 @@ autoware_msgs::ControlCommandStamped
 }
 
 void TwistFilter::configCallback(
-  const autoware_config_msgs::ConfigTwistFilterConstPtr& config)
+  const autoware_config_msgs::msg::ConfigTwistFilter::UniquePtr config)
 {
   lateral_accel_limit_ = config->lateral_accel_limit;
   lateral_jerk_limit_ = config->lateral_jerk_limit;
@@ -335,97 +366,51 @@ void TwistFilter::configCallback(
   lowpass_gain_steering_angle_ = config->lowpass_gain_steering_angle;
 }
 
-void TwistFilter::updatePrevTwist(const geometry_msgs::TwistStamped& msg)
+void TwistFilter::updatePrevTwist(const geometry_msgs::msg::TwistStamped& msg)
 {
   az_prev_.time = msg.header.stamp;
   az_prev_.val = msg.twist.angular.z;
 }
 
 void TwistFilter::updatePrevCtrl(
-  const autoware_msgs::ControlCommandStamped& msg)
+  const autoware_msgs::msg::ControlCommandStamped& msg)
 {
   sa_prev_.time = msg.header.stamp;
   sa_prev_.val = msg.cmd.steering_angle;
 }
 
 void TwistFilter::TwistCmdCallback(
-  const geometry_msgs::TwistStampedConstPtr& msg)
+  const geometry_msgs::msg::TwistStamped::UniquePtr msg)
 {
-  health_checker_.NODE_ACTIVATE();
   checkTwist(*msg);
-  geometry_msgs::TwistStamped ts;
+  geometry_msgs::msg::TwistStamped ts;
   ts = twist_filter::longitudinalLimitTwist(*msg, longitudinal_velocity_limit_);
   ts = _lon_accel_limiter.longitudinalAccelLimitTwist(ts);
   ts = lateralLimitTwist(ts);
   ts = smoothTwist(ts);
-  twist_pub_.publish(ts);
+  twist_pub_->publish(ts);
   publishLateralResultsWithTwist(ts);
   updatePrevTwist(ts);
 }
 
 void TwistFilter::CtrlCmdCallback(
-  const autoware_msgs::ControlCommandStampedConstPtr& msg)
+  const autoware_msgs::msg::ControlCommandStamped::UniquePtr msg)
 {
-  health_checker_.NODE_ACTIVATE();
   checkCtrl(*msg);
-  autoware_msgs::ControlCommandStamped ccs;
+  autoware_msgs::msg::ControlCommandStamped ccs;
   ccs = twist_filter::longitudinalLimitCtrl(*msg, longitudinal_velocity_limit_);
   ccs = _lon_accel_limiter.longitudinalAccelLimitCtrl(ccs);
   ccs = lateralLimitCtrl(ccs);
   ccs = smoothCtrl(ccs);
-  ctrl_pub_.publish(ccs);
+  ctrl_pub_->publish(ccs);
   publishLateralResultsWithCtrl(ccs);
   updatePrevCtrl(ccs);
 }
 
-TwistFilter::TwistFilter(ros::NodeHandle *nh, ros::NodeHandle *private_nh):
-    nh_(nh),
-    private_nh_(private_nh), 
-    health_checker_(*nh_, *private_nh_)
-{
-  if (nh_ != nullptr && private_nh_ != nullptr) {
-    nh_->param("vehicle_info/wheel_base", wheel_base_, 2.7);
-    private_nh_->param("longitudinal_velocity_limit", longitudinal_velocity_limit_, 80.0);
-    longitudinal_velocity_limit_ = longitudinal_velocity_limit_ * 0.44704;
-    private_nh_->param("longitudinal_accel_limit", longitudinal_accel_limit_, 3.5);
-    _lon_accel_limiter = LongitudinalAccelLimiter{
-      std::min(longitudinal_accel_limit_, 
-      MAX_LONGITUDINAL_ACCEL_HARDCODED_LIMIT_M_S_2)};
-    private_nh_->param("lateral_accel_limit", lateral_accel_limit_, 5.0);
-    private_nh_->param("lateral_jerk_limit", lateral_jerk_limit_, 5.0);
-    private_nh_->param("lowpass_gain_linear_x", lowpass_gain_linear_x_, 0.0);
-    private_nh_->param("lowpass_gain_angular_z", lowpass_gain_angular_z_, 0.0);
-    private_nh_->param(
-      "lowpass_gain_steering_angle", lowpass_gain_steering_angle_, 0.0);
 
-    twist_sub_ = nh_->subscribe(
-      "twist_raw", 1, &TwistFilter::TwistCmdCallback, this);
-    ctrl_sub_ = nh_->subscribe("ctrl_raw", 1, &TwistFilter::CtrlCmdCallback, this);
-    config_sub_ = nh_->subscribe(
-      "config/twist_filter", 10, &TwistFilter::configCallback, this);
+} // namespace twist_filter
 
-    twist_pub_ = nh_->advertise<geometry_msgs::TwistStamped>("twist_cmd", 5);
-    ctrl_pub_ =
-      nh_->advertise<autoware_msgs::ControlCommandStamped>("ctrl_cmd", 5);
-    twist_lacc_limit_debug_pub_ = private_nh_->advertise<std_msgs::Float32>(
-      "limitation_debug/twist/lateral_accel", 5);
-    twist_ljerk_limit_debug_pub_ = private_nh_->advertise<std_msgs::Float32>(
-      "limitation_debug/twist/lateral_jerk", 5);
-    ctrl_lacc_limit_debug_pub_ = private_nh_->advertise<std_msgs::Float32>(
-      "limitation_debug/ctrl/lateral_accel", 5);
-    ctrl_ljerk_limit_debug_pub_ = private_nh_->advertise<std_msgs::Float32>(
-      "limitation_debug/ctrl/lateral_jerk", 5);
-    twist_lacc_result_pub_ = private_nh_->advertise<std_msgs::Float32>(
-      "result/twist/lateral_accel", 5);
-    twist_ljerk_result_pub_ = private_nh_->advertise<std_msgs::Float32>(
-      "result/twist/lateral_jerk", 5);
-    ctrl_lacc_result_pub_ = private_nh_->advertise<std_msgs::Float32>(
-      "result/ctrl/lateral_accel", 5);
-    ctrl_ljerk_result_pub_ = private_nh_->advertise<std_msgs::Float32>(
-      "result/ctrl/lateral_jerk", 5);
-    health_checker_.ENABLE();
-  }
-}
+#include "rclcpp_components/register_node_macro.hpp"
 
-}  // namespace
-
+// Register the component with class_loader
+RCLCPP_COMPONENTS_REGISTER_NODE(twist_filter::TwistFilter)
