+++ core_planning/twist_filter/src/velocity_limit.cpp
@@ -1,5 +1,5 @@
 /*
- * Copyright (C) 2018-2021 LEIDOS.
+ * Copyright (C) 2022 LEIDOS.
  *
  * Licensed under the Apache License, Version 2.0 (the "License"); you may not
  * use this file except in compliance with the License. You may obtain a copy of
@@ -16,25 +16,25 @@
 
 #include "velocity_limit.hpp"
 
-namespace twist_filter {
-geometry_msgs::TwistStamped
-  longitudinalLimitTwist(const geometry_msgs::TwistStamped& msg, const double limit)
+namespace twist_filter{
+geometry_msgs::msg::TwistStamped
+  longitudinalLimitTwist(const geometry_msgs::msg::TwistStamped& msg, const double limit)
 {
-  geometry_msgs::TwistStamped ts;
+  geometry_msgs::msg::TwistStamped ts;
   ts = msg;
 
   auto orig_longitudinal_velocity = msg.twist.linear.x;
   auto longitudinal_velocity = msg.twist.linear.x;
 
   if (longitudinal_velocity > limit) {
-    ROS_WARN_STREAM("Longitudinal velocity of " << 
+    RCLCPP_WARN_STREAM(rclcpp::get_logger("twist_filter::TwistFilter"), "Longitudinal velocity of " << 
       orig_longitudinal_velocity << 
       " exceeds configured limit of " <<
       limit);
     longitudinal_velocity = limit;
   }
   if (longitudinal_velocity > MAX_LONGITUDINAL_VELOCITY_HARDCODED_LIMIT_M_S) {
-    ROS_ERROR_STREAM("Longitudinal velocity of " << 
+    RCLCPP_ERROR_STREAM(rclcpp::get_logger("twist_filter::TwistFilter"), "Longitudinal velocity of " << 
       orig_longitudinal_velocity << 
       " exceeds hard-coded limit of " <<
       MAX_LONGITUDINAL_VELOCITY_HARDCODED_LIMIT_M_S);
@@ -45,24 +45,24 @@ geometry_msgs::TwistStamped
   return ts;
 }
 
-autoware_msgs::ControlCommandStamped
-  longitudinalLimitCtrl(const autoware_msgs::ControlCommandStamped& msg, const double limit)
+autoware_msgs::msg::ControlCommandStamped
+  longitudinalLimitCtrl(const autoware_msgs::msg::ControlCommandStamped& msg, const double limit)
 {
-  autoware_msgs::ControlCommandStamped ccs;
+  autoware_msgs::msg::ControlCommandStamped ccs;
   ccs = msg;
 
   auto orig_longitudinal_velocity = msg.cmd.linear_velocity;
   auto longitudinal_velocity = msg.cmd.linear_velocity;
 
   if (longitudinal_velocity > limit) {
-    ROS_WARN_STREAM("Longitudinal velocity of " << 
+    RCLCPP_WARN_STREAM(rclcpp::get_logger("twist_filter::TwistFilter"),"Longitudinal velocity of " << 
       orig_longitudinal_velocity << 
       " exceeds configured limit of " <<
       limit);
     longitudinal_velocity = limit;
   }
   if (longitudinal_velocity > MAX_LONGITUDINAL_VELOCITY_HARDCODED_LIMIT_M_S) {
-    ROS_ERROR_STREAM("Longitudinal velocity of " << 
+    RCLCPP_ERROR_STREAM(rclcpp::get_logger("twist_filter::TwistFilter"), "Longitudinal velocity of " << 
       orig_longitudinal_velocity << 
       " exceeds hard-coded limit of " <<
       MAX_LONGITUDINAL_VELOCITY_HARDCODED_LIMIT_M_S);
@@ -72,4 +72,5 @@ autoware_msgs::ControlCommandStamped
   ccs.cmd.linear_velocity = longitudinal_velocity;
   return ccs;
 }
+
 }
