+++ core_planning/twist_gate/package.xml
@@ -6,19 +6,25 @@
   <maintainer email="h_ohta@ertl.jp">h_ohta</maintainer>
   <license>Apache 2</license>
 
-  <buildtool_depend>autoware_build_flags</buildtool_depend>
-  <buildtool_depend>catkin</buildtool_depend>
+  <buildtool_depend>ament_cmake</buildtool_depend>
+  <build_depend>carma_cmake_common</build_depend>
+  <build_depend>ament_auto_cmake</build_depend>
 
   <depend>autoware_config_msgs</depend>
-  <depend>autoware_health_checker</depend>
   <depend>autoware_msgs</depend>
   <depend>geometry_msgs</depend>
-  <depend>roscpp</depend>
-  <depend>rostest</depend>
-  <depend>rosunit</depend>
-  <depend>roslint</depend>
+  <depend>rclcpp_components</depend>
   <depend>std_msgs</depend>
-  <depend>tablet_socket_msgs</depend>
-  <depend>ros_observer</depend>
-  <build_depend>carma_cmake_common</build_depend>
+  <depend>carma_ros2_utils</depend>
+  <!-- <depend>tablet_socket_msgs</depend> -->
+
+    <test_depend>ament_lint_auto</test_depend>
+    <test_depend>ament_cmake_gtest</test_depend>
+
+    <exec_depend>launch</exec_depend>
+    <exec_depend>launch_ros</exec_depend>
+
+    <export>
+        <build_type>ament_cmake</build_type>
+    </export>
 </package>
