--- core_planning/twist_gate_ros2/package.xml
@@ -1,30 +1,0 @@
-<?xml version="1.0"?>
-<package format="2">
-  <name>twist_gate_ros2</name>
-  <version>1.12.0</version>
-  <description>The twist_gate package</description>
-  <maintainer email="h_ohta@ertl.jp">h_ohta</maintainer>
-  <license>Apache 2</license>
-
-  <buildtool_depend>ament_cmake</buildtool_depend>
-  <build_depend>carma_cmake_common</build_depend>
-  <build_depend>ament_auto_cmake</build_depend>
-
-  <depend>autoware_config_msgs</depend>
-  <depend>autoware_msgs</depend>
-  <depend>geometry_msgs</depend>
-  <depend>rclcpp_components</depend>
-  <depend>std_msgs</depend>
-  <depend>carma_ros2_utils</depend>
-  <!-- <depend>tablet_socket_msgs</depend> -->
-
-    <test_depend>ament_lint_auto</test_depend>
-    <test_depend>ament_cmake_gtest</test_depend>
-
-    <exec_depend>launch</exec_depend>
-    <exec_depend>launch_ros</exec_depend>
-
-    <export>
-        <build_type>ament_cmake</build_type>
-    </export>
-</package>
