--- core_planning/twist_gate_ros2/src/twist_gate_node.cpp
@@ -1,36 +1,0 @@
-/*
- * Copyright 2015-2019 Autoware Foundation. All rights reserved.
- *
- * Licensed under the Apache License, Version 2.0 (the "License");
- * you may not use this file except in compliance with the License.
- * You may obtain a copy of the License at
- *
- *     http://www.apache.org/licenses/LICENSE-2.0
- *
- * Unless required by applicable law or agreed to in writing, software
- * distributed under the License is distributed on an "AS IS" BASIS,
- * WITHOUT WARRANTIES OR CONDITIONS OF ANY KIND, either express or implied.
- * See the License for the specific language governing permissions and
- * limitations under the License.
- */
-
-// ROS Includes
-#include <rclcpp/rclcpp.hpp>
-
-// User defined includes
-#include <twist_gate/twist_gate.hpp>
-
-int main(int argc, char **argv) 
-{
-  rclcpp::init(argc, argv);
-
-  auto node = std::make_shared<TwistGate>(rclcpp::NodeOptions());
-  
-  rclcpp::executors::MultiThreadedExecutor executor;
-  executor.add_node(node->get_node_base_interface());
-  executor.spin();
-
-  rclcpp::shutdown();
-
-  return 0;
-}
